requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
bitarray==2.7.6
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.1.0
cytoolz==0.12.1
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.1.1
exceptiongroup==1.1.1
frozenlist==1.3.3
h11==0.14.0
hexbytes==0.3.1
httpcore==0.17.2
httpx==0.24.1
idna==3.4
importlib-resources==5.12.0
jsonschema==4.17.3
lru-dict==1.2.0
multidict==6.0.4
numpy==1.24.4
pandas==2.0.3
parsimonious==0.9.0
pkgutil_resolve_name==1.3.10
protobuf==4.23.3
pycryptodome==3.18.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-telegram-bot==13.6
pytz==2023.3
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
six==1.16.0
sniffio==1.3.0
toolz==0.12.0
tornado==6.3.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
web3==6.5.0
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
